this little hack resolves the issue with the home button navigation issue when offline
I'm pretty confident the issue is both the base path configuration and
how the caching is set up. The Home component links to "/". This gets
replaced by the base path config to be "/agua-vida-poc". But the
cache logic is explicitly looking for "/agua-vida-poc/" so it can
return 'index.html'. Since it can't match, it gives us that page not
found error when offline. Jesus Christ.